Replace hooking mechanism, modify module structures, rewrite message resources
Remove HookBase and Hook classes in MetaTweetServer, and replace to the new
hooking mechanism in LinxFramework.
Since hooking mechanism is replaced, modify and add some methods in the module
classes, consumer of the hooks.
Remove Initialize(XmlConfiguration) method in IModule interface and modify the
signature of IModule#Register method; initialization of configuration file for
modules operated by (default) Initializer class is now in Register method, not
in Initialize(XmlConfiguration) method.
Rename the methods to initialize ObjectContext not to shadow Initialize method
in IModule for StorageModule.
Clean up and rewrite the string message resources in MetaTweetServer.