Add a method definition in IModule, modify key data type of CodeDomain
Add the definition of CreateObjRef(Type) method in IModule, that practically
limits the base class of IModule-implemented-classes to MarshalByRefObject
because the ModuleDomain have to know the type information without loading
module assemblies.
Modify the type of key for IModule in CodeDomain to String instead of Type
since Type causes assembly loading.